Created new version of reporter.py
This new version fixes issues with apple ids that have multiple accounts.
Before, it was not possible to request getVendors() and getReports() etc. for apple ids with multiple accounts because the account number was passed as a=account number. This is no longer valid syntax. Instead, apple wants the account number to be part of Reporter.properties. Equivalent to this, one can pass the account number in the data argument.